requirements.txt
streamlit
openai
llama_index
weaviate-client
